+++ src/main/java/chatter/TaskList.java
@@ -1,5 +1,6 @@
-import java.io.FileWriter;
-import java.io.IOException;
+package chatter;
+
+import chatter.task.Task;
 
 import java.util.ArrayList;
 
@@ -14,9 +15,9 @@ public class TaskList {
     private int numOfTasks = 0;
 
     /**
-     * A constructor to initialize the TaskList class.
+     * A constructor to initialize the chatter.TaskList class.
      *
-     * @param list ArrayList of Task objects from reading local data file.
+     * @param list ArrayList of chatter.task.Task objects from reading local data file.
      */
     public TaskList(ArrayList<Task> list) {
         this.list = list;
@@ -24,16 +25,16 @@ public class TaskList {
     }
 
     /**
-     * A constructor to initialize the TaskList class.
+     * A constructor to initialize the chatter.TaskList class.
      */
     public TaskList() {
         this.list = new ArrayList<>();
     }
 
     /**
-     * Adds a Task object to the list of Tasks.
+     * Adds a chatter.task.Task object to the list of Tasks.
      *
-     * @param task The Task object to be added.
+     * @param task The chatter.task.Task object to be added.
      */
     public void addTask(Task task) {
         list.add(task);
@@ -41,7 +42,7 @@ public class TaskList {
     }
 
     /**
-     * Marks specific Task in list as completed.
+     * Marks specific chatter.task.Task in list as completed.
      *
      * @param taskNumber Index number of task in list to be mark as completed.
      */
@@ -51,7 +52,7 @@ public class TaskList {
     }
 
     /**
-     * Deletes specific Task in list.
+     * Deletes specific chatter.task.Task in list.
      *
      * @param taskNumber Number of task in list to be deleted.
      */
@@ -61,7 +62,7 @@ public class TaskList {
     }
 
     /**
-     * Marks specific Task in list as uncompleted.
+     * Marks specific chatter.task.Task in list as uncompleted.
      *
      * @param taskNumber Number of task in list to be mark as uncompleted.
      */
@@ -80,9 +81,9 @@ public class TaskList {
     }
 
     /**
-     * Returns a string of the list of Task objects in storage string format.
+     * Returns a string of the list of chatter.task.Task objects in storage string format.
      *
-     * @return String of list of Task objects.
+     * @return String of list of chatter.task.Task objects.
      */
     public String toStorageString() {
         StringBuilder storageString = new StringBuilder();
@@ -96,9 +97,9 @@ public class TaskList {
     }
 
     /**
-     * Returns a specific Task object at the specified index.
+     * Returns a specific chatter.task.Task object at the specified index.
      *
-     * @return Task object at the specified index.
+     * @return chatter.task.Task object at the specified index.
      */
     public Task getTask(int index) {
         return list.get(index);
