+++ src/main/java/chatter/command/Command.java
@@ -1,5 +1,12 @@
+package chatter.command;
+
+import chatter.ChatterException;
+import chatter.Storage;
+import chatter.TaskList;
+import chatter.Ui;
+
 /**
- * Represents an abstract Command class.
+ * Represents an abstract chatter.command.Command class.
  *
  * @author Anthony Tamzil
  * @version CS2103T Individual Project AY2023/24 Semester 1
@@ -9,7 +16,7 @@ public abstract class Command {
     private boolean isExit;
 
     /**
-     * Constructor for Command class.
+     * Constructor for chatter.command.Command class.
      */
     public Command() {
         this.isExit = false;
@@ -35,9 +42,9 @@ public abstract class Command {
     /**
      * Abstract method that executes the command
      *
-     * @param tasks TaskList class storing an ArrayList of Task objects.
-     * @param ui Ui class to handle user interactions.
-     * @param storage Storage class to read and store tasks by the user.
+     * @param tasks chatter.TaskList class storing an ArrayList of chatter.task.Task objects.
+     * @param ui chatter.Ui class to handle user interactions.
+     * @param storage chatter.Storage class to read and store tasks by the user.
      * @throws ChatterException Any error that occurs due to the command.
      */
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws ChatterException;
