+++ src/main/java/chatter/command/MarkCommand.java
@@ -1,18 +1,24 @@
+package chatter.command;
+
+import chatter.Storage;
+import chatter.TaskList;
+import chatter.Ui;
+
 /**
- * Represents a Command class that is responsible for marking / unmarking a task as done.
+ * Represents a chatter.command.Command class that is responsible for marking / unmarking a task as done.
  *
  * @author Anthony Tamzil
  * @version CS2103T Individual Project AY2023/24 Semester 1
  */
-public class MarkCommand extends Command{
-    /** Task number to be marked as done. */
+public class MarkCommand extends Command {
+    /** chatter.task.Task number to be marked as done. */
     private int taskNumber;
     /** Boolean containing whether user wants to mark or unmark a task as done. */
     private boolean isDone;
 
 
     /**
-     * Constructor for the MarkCommand class for users to mark a task as done.
+     * Constructor for the chatter.command.MarkCommand class for users to mark a task as done.
      */
     public MarkCommand(boolean isDone, int taskNumber) {
         this.taskNumber = taskNumber;
@@ -22,9 +28,9 @@ public class MarkCommand extends Command{
     /**
      * Marks the specified task as done.
      *
-     * @param tasks TaskList class storing an ArrayList of Task objects.
-     * @param ui Ui class to handle user interactions.
-     * @param storage Storage class to read and store tasks by the user.
+     * @param tasks chatter.TaskList class storing an ArrayList of chatter.task.Task objects.
+     * @param ui chatter.Ui class to handle user interactions.
+     * @param storage chatter.Storage class to read and store tasks by the user.
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
