+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 -----------------------
-Hello! I'm Chatter!
+Hello! I'm chatter.Chatter!
 How can i help you today?
 -----------------------
 -----------------------
